refactor:

+++ Models/PlayerSaves/PlayerInfo.cs
@@ -2,16 +2,30 @@
 
 public record PlayerInfo
 {
-    public long Pid { get; init; }
+    /// <summary>
+    /// player id
+    /// </summary>
+    public int Pid { get; init; }
+    /// <summary>
+    /// player name
+    /// </summary>
     public string Name { get; init; }
+    /// <summary>
+    /// player avatar
+    /// </summary>
     public string Pic { get; init; }
-    public long Cash { get; init; }
+
+    public int Cash { get; init; }
     public int CompletedTutorial { get; init; }
+
     public int DefaultMap { get; init; }
     public List<string> MapNames { get; init; }
     public List<int> MapSizes { get; init; }
+
     public int WorldId { get; init; }
+
     public int SpRefUid { get; init; }
     public string SpRefCatInstall { get; init; }
+
     public long LastLoggedIn { get; init; }
 }
